% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{admm_genlasso_PQ}
\alias{admm_genlasso_PQ}
\title{ADMM optimizer for generalized lasso problems}
\usage{
admm_genlasso_PQ(
  y,
  X,
  w,
  nu,
  pL2,
  beta0,
  lambda,
  rho,
  precondition,
  objtol,
  reltol,
  abstol,
  maxiter,
  history
)
}
\arguments{
\item{y}{(arma::vec) response vector.}

\item{w}{(arma::vec) weight vector.}

\item{nu}{(arma::vec) penalty weight vector.}

\item{pL2}{(arma::vec) ridge penalty vector.}

\item{beta0}{(arma::vec) initial values.}

\item{lambda}{(double) generalized lasso penalty parameter.}

\item{rho}{(double) augmented Lagrangian penalty parameter.}

\item{objtol}{(double) objective function tollerance parameter.}

\item{reltol}{(double) relative tollerance parameter.}

\item{abstol}{(double) absolute tollerance parameter.}

\item{maxiter}{(int) maximum numner of iterations.}

\item{x}{(arma::mat) design matrix.}
}
\value{
A list containing the estimated coefficients and the optimization history.
}
\description{
ADMM optimizer for generalized lasso problems
}
