% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chol_solve}
\alias{chol_solve}
\alias{chol_rnd_sim}
\alias{chol_logdet}
\alias{chol_edf}
\title{Cholesky solvers for dense matrix operations}
\usage{
chol_solve(R, r)

chol_rnd_sim(R, r)

chol_logdet(R)

chol_edf(R, B, exact = FALSE, rank = 10, nsample = 100)
}
\arguments{
\item{R}{upper-triangular factor of the Cholesky decomposition}

\item{r}{vector or matrix on the right-hand side of the linear system}

\item{B}{cross-product matrix}
}
\description{
\itemize{
\item \code{chol_solve}: solve a linear system of equations
\item \code{chol_rnd_sim}: generate a Gaussian random vector
\item \code{chol_logdet}: log-determinant of a matrix
\item \code{chol_edf}: Hutch++ estimator of the EDF
}
}
\keyword{internal}
