% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coeff_PQ}
\alias{coeff_PQ}
\title{PQ bound coefficients
Remark: correcting for numerical instability around 0
numerical instability still present at Infinity}
\usage{
coeff_PQ(eta)
}
\arguments{
\item{eta}{(arma::vec(n)) linear predictors}
}
\value{
PQ w and nu (as an arma::vec(2*n))
}
\description{
PQ bound coefficients
Remark: correcting for numerical instability around 0
numerical instability still present at Infinity
}
