% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logit_enet_coord.R
\name{fit_logit_enet_coord}
\alias{fit_logit_enet_coord}
\title{Ridge logistic regression via MM optimization}
\usage{
fit_logit_enet_coord(
  y,
  X,
  type = c("NR", "BL", "PG", "PQ"),
  beta_start = NULL,
  lambda = NULL,
  alpha = 0.99,
  eps = 1e-10,
  intercept = FALSE,
  phi = 0.9,
  maxiter = 1000L,
  abstol = 1e-04,
  reltol = 1e-04,
  etatol = 0.1,
  verbose = FALSE,
  freq = 10L
)
}
\arguments{
\item{y}{A \verb{n} dimensional binary vector}

\item{type}{(string) Family of surrogate functions for iterative optimization. Must be one of 'NR', 'BL', 'PG', or 'PQ'}

\item{beta_start}{(real vector) Initialization point. If missing, the algorithm is initialized setting at zero all the coefficients}

\item{lambda}{(real) Regularization parameters for ridge regression}

\item{alpha}{(real) Mixing parameter regulating the balancing between lasso and ridge penalties.}

\item{eps}{(real) Small value added to the intercept term to avoid numerical issues}

\item{intercept}{(boolean) Is the intercept included in \code{X} or not?}

\item{phi}{(real) inertia parameter for the active set detection}

\item{maxiter}{(integer) Maximum number of iterations. If reached, the algorithm raises an error}

\item{abstol}{(real) Convergence threshold for the absolute change in the log-likelihood}

\item{reltol}{(real) Convergence threshold for the relative change in the log-likelihood}

\item{etatol}{(real) Convergence threshold for the relative change in the linear predictor}

\item{verbose}{(boolean) Print the intermediate state of the optimization}

\item{freq}{(int) How often print the optimization state}

\item{x}{A \verb{n x p} dimensional design matrix. The intercept must be manually included}

\item{use_nn}{(boolean) Use nearest neighbors search as safety check on PQ bounds optimization in later iterations.}
}
\value{
The function returns a list with optimal coefficients and likelihood path through MM iterations
}
\description{
Ridge logistic regression via MM optimization
}
