% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logit_mfvb_adj.R
\name{fit_logit_mcmc_adj}
\alias{fit_logit_mcmc_adj}
\title{Ridge logistic regression via PG Gibbs sampling}
\usage{
fit_logit_mcmc_adj(
  y,
  X,
  D = diag(ncol(X)),
  beta_start = NULL,
  lambda = 1,
  eps = 1e-10,
  intercept = FALSE,
  hpar = list(A = 1, B = 1, v = 3),
  prior = c("auto", "IG", "HW"),
  solver = c("auto", "chol", "smw", "sparse"),
  scaling = TRUE,
  nscaling = 10,
  maxiter = 5000L,
  burn = floor(maxiter/2),
  thin = 5L,
  verbose = FALSE,
  freq = floor(maxiter/20),
  seed = 1234
)
}
\arguments{
\item{y}{A \verb{n} dimensional binary vector}

\item{X}{A \verb{n x p} dimensional design matrix. The intercept must be manually included}

\item{D}{A \verb{m x p} dimensional penalty matrix.}

\item{beta_start}{(real vector) Initialization point. If missing, the algorithm is initialized setting at zero all the coefficients}

\item{lambda}{(real) Regularization parameters for ridge regression}

\item{eps}{(real) Small value added to the intercept term to avoid numerical issues}

\item{intercept}{(boolean) Is the intercept included in \code{X} or not?}

\item{solver}{(string) Solver used for the linear systems and implicit matrix inversion. Must be one of 'auto', 'chol', 'smw', or 'sparse'}

\item{maxiter}{(integer) Maximum number of iterations. If reached, the algorithm raises an error}

\item{burn}{(integer) Burn-in period.}

\item{thin}{(integer) Thinning frequency.}

\item{verbose}{(boolean) Print the intermediate state of the optimization}

\item{freq}{(int) How often print the optimization state}
}
\value{
The function returns a list with optimal coefficients and likelihood path through MM iterations
}
\description{
Ridge logistic regression via PG Gibbs sampling
}
