% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logit_mfvb.R
\name{fit_logit_mfvb}
\alias{fit_logit_mfvb}
\alias{fit_logit_pqvb}
\alias{fit_logit_pgvb}
\alias{fit_logit_blvb}
\title{Ridge logistic regression via variational approximation}
\usage{
fit_logit_mfvb(
  y,
  X,
  D = diag(ncol(X)),
  type = c("PQ", "PG", "BL"),
  beta_start = NULL,
  lambda = 1,
  eps = 1e-10,
  intercept = FALSE,
  solver = c("auto", "chol", "smw", "sparse"),
  maxiter = 1000L,
  abstol = 1e-04,
  reltol = 1e-04,
  verbose = FALSE,
  freq = 10L
)

fit_logit_pqvb(
  y,
  X,
  D = diag(ncol(X)),
  beta_start = NULL,
  lambda = 1,
  eps = 1e-10,
  intercept = FALSE,
  solver = c("auto", "chol", "smw", "sparse"),
  maxiter = 1000L,
  abstol = 1e-04,
  reltol = 1e-04,
  verbose = FALSE,
  freq = 10L
)

fit_logit_pgvb(
  y,
  X,
  D = diag(ncol(X)),
  beta_start = NULL,
  lambda = 1,
  eps = 1e-10,
  intercept = FALSE,
  solver = c("auto", "chol", "smw", "sparse"),
  maxiter = 1000L,
  abstol = 1e-04,
  reltol = 1e-04,
  verbose = FALSE,
  freq = 10L
)

fit_logit_blvb(
  y,
  X,
  D = diag(ncol(X)),
  beta_start = NULL,
  lambda = 1,
  eps = 1e-10,
  intercept = FALSE,
  solver = c("auto", "chol", "smw", "sparse"),
  maxiter = 1000L,
  abstol = 1e-04,
  reltol = 1e-04,
  verbose = FALSE,
  freq = 10L
)
}
\arguments{
\item{y}{A \verb{n} dimensional binary vector}

\item{X}{A \verb{n x p} dimensional design matrix. The intercept must be manually included}

\item{D}{A \verb{m x p} dimensional penalty matrix.}

\item{type}{(string) Family of surrogate functions for iterative optimization. Must be one of 'BL', 'PG', or 'PQ'}

\item{beta_start}{(real vector) Initialization point. If missing, the algorithm is initialized setting at zero all the coefficients}

\item{lambda}{(real) Regularization parameters for ridge regression}

\item{eps}{(real) Small value added to the intercept term to avoid numerical issues}

\item{intercept}{(boolean) Is the intercept included in \code{X} or not?}

\item{solver}{(string) Solver used for the linear systems and implicit matrix inversion. Must be one of 'auto', 'chol', 'smw', or 'sparse'}

\item{maxiter}{(integer) Maximum number of iterations. If reached, the algorithm raises an error}

\item{abstol}{(real) Convergence threshold for the absolute change in the log-likelihood}

\item{reltol}{(real) Convergence threshold for the relative change in the log-likelihood}

\item{verbose}{(boolean) Print the intermediate state of the optimization}

\item{freq}{(int) How often print the optimization state}
}
\value{
The function returns a list with optimal coefficients and likelihood path through MM iterations
}
\description{
Ridge logistic regression via variational approximation
}
