% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logit_genlasso.R
\name{fit_logit_splasso}
\alias{fit_logit_splasso}
\title{Generalized lasso logistic regression via MM optimization}
\usage{
fit_logit_splasso(
  y,
  X,
  D,
  type = c("NR", "BL", "PG", "PQ"),
  beta_start = NULL,
  lambda = 1,
  alpha = 0.99,
  eps = 1e-10,
  phi = 0.9,
  intercept = FALSE,
  prox = TRUE,
  maxiter = 1000,
  abstol = 1e-04,
  reltol = 1e-04,
  etatol = 1e-04,
  verbose = FALSE,
  freq = 10,
  ctr_admm = set_ctr_admm()
)
}
\arguments{
\item{y}{A \verb{n} dimensional binary vector.}

\item{X}{A \verb{n x p} dimensional design matrix. The intercept must be manually included.}

\item{D}{A \verb{m x p} dimensional penalty matrix.}

\item{type}{(string) Family of surrogate functions for iterative optimization. Must be one of 'NR', 'BL', 'PG', or 'PQ'}

\item{beta_start}{(real vector) Initialization point. If missing, the algorithm is initialized setting at zero all the coefficients.}

\item{lambda}{(real) Regularization parameter for elastic-net regression.}

\item{alpha}{(real) Mixing parameter regulating the balancing between lasso and ridge penalties.}

\item{eps}{(real) Small value added to the intercept term to avoid numerical issues.}

\item{phi}{(real) Relaxation parameter for the dual variables}

\item{intercept}{(boolean) Is the intercept included in \code{X} or not?}

\item{prox}{(boolean) Use the proximal primal-dual update}

\item{maxiter}{(integer) Maximum number of iterations. If reached, the algorithm raises an error.}

\item{abstol}{(real) Convergence threshold for the absolute change in the log-likelihood.}

\item{reltol}{(real) Convergence threshold for the relative change in the log-likelihood.}

\item{etatol}{(real) Convergence threshold for the relative change in the linear predictor.}

\item{verbose}{(boolean) Print the intermediate state of the optimization.}

\item{freq}{(int) How often print the optimization state.}

\item{ctr_admm}{(list) Control parameters for the ADMM algorithm}
}
\value{
The function returns a list with optimal coefficients and likelihood path through MM iterations
}
\description{
Generalized lasso logistic regression via MM optimization
}
