% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logit_ridge.R
\name{fit_logit_spridge}
\alias{fit_logit_spridge}
\title{Ridge logistic regression via MM optimization}
\usage{
fit_logit_spridge(
  y,
  X,
  D,
  type = c("NR", "BL", "PG", "PQ"),
  beta_start = NULL,
  lambda = NULL,
  eps = 1e-10,
  intercept = FALSE,
  phi = 0.9,
  maxiter = 1000L,
  abstol = 1e-04,
  reltol = 1e-04,
  etatol = 1e-04,
  verbose = FALSE,
  freq = 10L,
  use_nn = FALSE,
  method = c("prjg", "dual", "admm"),
  ctr_admm = set_ctr_admm(),
  ctr_prjg = set_ctr_prjg(),
  ctr_dual = set_ctr_dual()
)
}
\arguments{
\item{y}{A \verb{n} dimensional binary vector}

\item{D}{A \verb{q x p} dimensional penalty matrix. The intercept must be manually included}

\item{type}{(string) Family of surrogate functions for iterative optimization. Must be one of 'NR', 'BL', 'PG', or 'PQ'}

\item{beta_start}{(real vector) Initialization point. If missing, the algorithm is initialized setting at zero all the coefficients}

\item{lambda}{(real) Regularization parameters for ridge regression}

\item{eps}{(real) Small value added to the intercept term to avoid numerical issues}

\item{intercept}{(boolean) Is the intercept included in \code{X} or not?}

\item{maxiter}{(integer) Maximum number of iterations. If reached, the algorithm raises an error}

\item{abstol}{(real) Convergence threshold for the absolute change in the log-likelihood}

\item{reltol}{(real) Convergence threshold for the relative change in the log-likelihood}

\item{etatol}{(real) Convergence threshold for the relative change in the linear predictor}

\item{verbose}{(boolean) Print the intermediate state of the optimization}

\item{freq}{(int) How often print the optimization state}

\item{use_nn}{(boolean) Use nearest neighbors search as safety check on PQ bounds optimization in later iterations.}

\item{method}{(string) Method to be used for the solution of the quadratic programming inner optimization. Must be one of 'dual', or 'admm'}

\item{ctr_admm}{(list) Control parameters for the ADMM algorithm}

\item{ctr_prjg}{(list) Control parameters for the projected gradient algorithm}

\item{ctr_dual}{(list) Control parameters for the dual QP algorithm}

\item{x}{A \verb{n x p} dimensional design matrix. The intercept must be manually included}
}
\value{
The function returns a list with optimal coefficients and likelihood path through MM iterations
}
\description{
Ridge logistic regression via MM optimization
}
