% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_genlasso.R
\name{init_splasso_beta}
\alias{init_splasso_beta}
\title{Initialize the regression parameters for spatial-lasso logistic regression}
\usage{
init_splasso_beta(y, X, D, pL1, pL2, spthr = 0.9)
}
\description{
Initialize the regression parameters for spatial-lasso logistic regression
}
\keyword{internal}
