% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticPQbound.R
\name{invQ_r}
\alias{invQ_r}
\title{accurate inverse-matrix vector product}
\usage{
invQ_r(cholQ, r)
}
\arguments{
\item{cholQ}{(real matrix) Cholesky matrix}

\item{r}{(real vector) vector to be pre-multiplied with inverse}
}
\description{
accurate inverse-matrix vector product
}
