% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invQ_r_Rcpp}
\alias{invQ_r_Rcpp}
\title{accurate inverse-matrix vector product Q^{-1}r via Cholesky factorization}
\usage{
invQ_r_Rcpp(cholQ, r)
}
\arguments{
\item{cholQ}{the Cholesky factor of Q}

\item{r}{A vector}
}
\description{
accurate inverse-matrix vector product Q^{-1}r via Cholesky factorization
}
