% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitPQbound-package.R
\docType{package}
\name{logitPQbound-package}
\alias{logitPQbound}
\alias{logitPQbound-package}
\title{logitPQbound: Implementation of the Novel Piece-Wise Quadratic Bound for Logistic Log-Likelihoods}
\description{
Reproduction of the results from the main manuscript.
}
\keyword{internal}
