% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logpost}
\alias{logpost}
\title{logistic penalized log likelihood under elastic net regularization}
\usage{
logpost(eta, y, beta, alpha, lambda)
}
\arguments{
\item{eta}{(arma::vec(n)) linear predictors}

\item{y}{(arma::vec(n)) responses}

\item{beta}{(arma::vec(p)) parameters (including intercept)}

\item{alpha}{(double) relative weight of L1 and L2 penalty}

\item{lambda}{(double) regularization parameter}
}
\value{
logistic penalized log likelihood
}
\description{
logistic penalized log likelihood under elastic net regularization
}
