% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticPQbound.R
\name{logpost_R}
\alias{logpost_R}
\title{logistic penalized log likelihood under elastic net regularization}
\usage{
logpost_R(eta, y, beta, alpha, lambda)
}
\arguments{
\item{eta}{(real vector) linear predictors}

\item{y}{(binary vector) responses}

\item{beta}{(real vector) parameters (including intercept)}

\item{alpha}{(real) relative weight of L1 and L2 penalty}

\item{lambda}{(real) regularization parameter}
}
\value{
logistic penalized log likelihood
}
\description{
logistic penalized log likelihood under elastic net regularization
}
