% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{maxreldiff}
\alias{maxreldiff}
\title{Maximum absolute relative difference operator}
\usage{
maxreldiff(xnew, xold)
}
\arguments{
\item{xnew}{new scalar, vector or matrix value}

\item{xold}{old scalar, vector or matrix value}
}
\description{
Maximum absolute relative difference operator
}
\keyword{internal}
