% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prjgrad_large_n}
\alias{prjgrad_large_n}
\title{projected gradient descent, optimized for large-n-small-p problems}
\usage{
prjgrad_large_n(u, y_nu, nu_x, cholQ, alpha, thrs, maxiter)
}
\arguments{
\item{u}{(arma::vec(n)) initial solution}

\item{y_nu}{(arma::vec(n)) auxiliary vector}

\item{nu_x}{(arma::mat(n,p)) auxiliary matrix}

\item{cholQ}{(arma::mat(p,p)) auxiliary Cholesky matrix}

\item{alpha}{(double) gradient step size}

\item{thrs}{(double) stopping threshold on relative absolute change}

\item{maxiter}{(int) maximum number of iterations}
}
\value{
logistic penalized log likelihood
}
\description{
projected gradient descent, optimized for large-n-small-p problems
}
