% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticPQbound.R
\name{ridge_logit}
\alias{ridge_logit}
\title{Ridge logistic regression via MM optimization}
\usage{
ridge_logit(
  x,
  y,
  type = "PQ",
  beta_start = NULL,
  lambda = NULL,
  maxiter = 10000,
  tol = 1e-10,
  eps_pen = 1e-10,
  maxiter_in = 100,
  tol_in = 0.1,
  alpha_prjg = 0.1,
  use_qp = F,
  use_nn = F
)
}
\arguments{
\item{x}{A \verb{n x p} dimensional design matrix. The intercept must be manually included.}

\item{y}{A \verb{n} dimensional binary vector.}

\item{type}{(string) Family of surrogate functions for iterative optimization. Must be one of 'NR', 'BL', 'PG', or 'PQ'}

\item{beta_start}{(real vector) Initialization point. If missing, the algorithm is initialized setting at zero all the coefficients.}

\item{lambda}{(real) Regularization parameters for ridge regression}

\item{maxiter}{(integer) Maximum number of iterations. If reached, the algorithm raises an error.}

\item{tol}{(real) Threshold at which the algorithm stops.}

\item{eps_pen}{(real) Small value added to the diagonal of the Hessian matrix to avoid numerical issues.}

\item{maxiter_in}{(integer) Maximum number of inner iterations for PQ bound optimization (either for projected gradient or quadratic programming).}

\item{tol_in}{(real) Threshold at which the inner optimization of PQ bound stops (either for projected gradient or quadratic programming).}

\item{alpha_prjg}{(real) Step size for projected gradient descent on PQ bounds.}

\item{use_qp}{(boolean) Use quadratic programming inner optimization of PQ bounds.}

\item{use_nn}{(boolean) Use nearest neighbors search as safety check on PQ bounds optimization in later iterations.}
}
\value{
The function returns a list with optimal coefficents and likelihood path through MM iterations
}
\description{
Ridge logistic regression via MM optimization
}
