% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{set_ctr_admm}
\alias{set_ctr_admm}
\title{Set the control parameters of the ADMM algorithm}
\usage{
set_ctr_admm(
  rho = 1,
  tau = 0,
  gamma = 0,
  intercept = FALSE,
  spthr = 0.9,
  smw = FALSE,
  precondition = FALSE,
  objtol = 0.001,
  reltol = 0.001,
  abstol = 1e-04,
  maxiter = 100,
  verbose = FALSE,
  freq = 10
)
}
\arguments{
\item{rho}{Penalty parameter.}

\item{tau}{Relaxation parameter.}

\item{gamma}{Proximal parameter.}

\item{intercept}{Is the intercept included in \code{X} or not?}

\item{spthr}{Percentage of non-zero element in the system matrix to activate sparse solvers.}

\item{smw}{Must SMW formula be used?}

\item{precondition}{Must Jacobi preconditioning be used?}

\item{objtol}{Convergence threshold for the absolute change in the log-likelihood.}

\item{reltol}{Convergence threshold for the relative change in the primal and dual variables.}

\item{abstol}{Convergence threshold for the absolute change in the the primal and dual variables.}

\item{maxiter}{Maximum number of iterations. If reached.}

\item{verbose}{Print the intermediate state of the optimization.}

\item{freq}{How often print the optimization state.}
}
\description{
Set the control parameters of the ADMM algorithm
}
