% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{set_ctr_dual}
\alias{set_ctr_dual}
\title{Set the control parameters of the dual QP algorithm}
\usage{
set_ctr_dual(tol = 1e-04, maxiter = 100)
}
\description{
Set the control parameters of the dual QP algorithm
}
\keyword{internal}
