% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{set_ctr_prjg}
\alias{set_ctr_prjg}
\title{Set the control parameters of the projected gradient algorithm}
\usage{
set_ctr_prjg(stepsize = 0.1, tol = 0.01, maxiter = 100)
}
\description{
Set the control parameters of the projected gradient algorithm
}
\keyword{internal}
