% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smw_chol_solve}
\alias{smw_chol_solve}
\alias{smw_chol_logdet}
\alias{smw_chol_edf}
\title{SMW-Cholesky solvers for dense matrix operations}
\usage{
smw_chol_solve(R, X, pL2, r)

smw_chol_logdet(R, w, pL2)

smw_chol_edf(R, B, w, pL2)
}
\arguments{
\item{R}{upper-triangular factor of the Cholesky decomposition}

\item{X}{design matrix}

\item{pL2}{penalty vector}

\item{r}{vector or matrix on the right-hand side of the linear system}

\item{w}{weight vector}

\item{B}{cross-product matrix}
}
\description{
\itemize{
\item \code{smw_chol_solve}: solve a linear system of equations
\item \code{smw_chol_rnd_sim}: generate a Gaussian random vector
\item \code{smw_chol_logdet}: log-determinant of a matrix
\item \code{smw_chol_edf}: Hutch++ estimator of the EDF
}
}
\keyword{internal}
