% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smw_solve}
\alias{smw_solve}
\alias{smw_rnd_sim}
\alias{smw_logdet}
\alias{smw_edf}
\title{SMW solvers for dense matrix operations}
\usage{
smw_solve(X, w, pL2, r)

smw_rnd_sim(X, y, pL2)

smw_logdet(X, w, pL2)

smw_edf(X, w, pL2)
}
\arguments{
\item{X}{design matrix}

\item{w}{weight vector}

\item{pL2}{penalty vector}

\item{r}{vector or matrix on the right-hand side of the linear system}

\item{y}{response vector}
}
\description{
\itemize{
\item \code{smw_solve}: solve a linear system of equations
\item \code{smw_rnd_sim}: generate a Gaussian random vector
\item \code{smw_logdet}: log-determinant of a matrix
\item \code{smw_edf}: Hutch++ estimator of the EDF
}
}
\keyword{internal}
