% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{soft_sign}
\alias{soft_sign}
\title{Element-wise soft-sign operator}
\usage{
soft_sign(x, thr)
}
\arguments{
\item{x}{vector or matrix to be thresholded}

\item{thr}{thresholding parameter}
}
\description{
Element-wise soft-sign operator
}
\keyword{internal}
