% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{soft_threshold}
\alias{soft_threshold}
\title{Element-wise soft-thresholding operator}
\usage{
soft_threshold(x, lambda)
}
\arguments{
\item{x}{vector or matrix to be thresholded}

\item{lambda}{thresholding parameter}
}
\description{
Element-wise soft-thresholding operator
}
\keyword{internal}
