% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_genlasso.R
\name{solve_admm_genenet_PQ}
\alias{solve_admm_genenet_PQ}
\title{Compute the generalized elastic-net solution using the ADMM algorithm}
\usage{
solve_admm_genenet_PQ(
  X,
  y,
  w,
  nu,
  D = diag(ncol(X)),
  d = rep(0, nrow(D)),
  beta0 = NULL,
  z0 = NULL,
  u0 = NULL,
  lambda = 1,
  alpha = 0.99,
  eps = 1e-08,
  rho = 1,
  tau = 0,
  gamma = 0,
  intercept = TRUE,
  spthr = 0.9,
  smw = FALSE,
  precondition = FALSE,
  objtol = 0.001,
  reltol = 0.01,
  abstol = 1e-04,
  maxiter = 1000,
  verbose = FALSE,
  freq = 10
)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Response vector}

\item{w}{L2 Weight vector}

\item{nu}{L1 weight vector}

\item{D}{Penalty matrix}

\item{beta0}{Optional initial value}

\item{lambda}{Penalty parameter}

\item{alpha}{Elastic-net mixing parameter}

\item{eps}{Intercept specific penalty parameter}

\item{rho}{Augmented Lagrangian penalty parameter}

\item{tau}{Dual over-relaxation parameter}

\item{gamma}{Proximal regularization parameter}

\item{intercept}{If \code{TRUE}, do not penalize per first column of \code{X}}

\item{spthr}{Proportion of zero entries to transform \code{X} and \code{D} into sparse matrices}

\item{smw}{If \code{TRUE}, allows for the SMW solver in the \code{p>n} regime}

\item{reltol}{Relative tolerance stopping criterion}

\item{abstol}{Absolute tolerance stopping criterion}

\item{maxiter}{Maximum number of iterations}
}
\value{
\itemize{
\item \code{beta}: estimated primal coefficients
\item \code{z}: estimated primal variable
\item \code{u}: estimated dual variables
\item \code{r}: primal residuals
\item \code{s}: dual residuals
\item \code{niter}: number of iterations to reach convergence
\item \code{trace}: data-frame collecting the optimization history
}
}
\description{
Compute the generalized elastic-net solution using the ADMM algorithm
}
