% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_genlasso.R
\name{solve_admm_ridge_PQ}
\alias{solve_admm_ridge_PQ}
\title{ADMM optimizer for the PQ generalized lasso problem}
\usage{
solve_admm_ridge_PQ(
  X,
  y,
  w,
  nu,
  pL2,
  beta0 = NULL,
  rho = 1,
  precondition = FALSE,
  objtol = 1e-05,
  reltol = 0.01,
  abstol = 1e-04,
  maxiter = 1000,
  history = FALSE
)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{w}{weight vector}

\item{nu}{penalty weight vector}

\item{pL2}{ridge penalty vector}

\item{beta0}{initial values}

\item{rho}{augmented Lagrangian penalty parameter}

\item{objtol}{objective tolerance parameter}

\item{reltol}{relative tolerance parameter}

\item{abstol}{absolute tolerance parameter}

\item{maxiter}{maximum number of iterations}

\item{lambda}{generalized lasso penalty parameter}
}
\value{
A list containing the estimated coefficients and the optimization history.
}
\description{
ADMM optimizer for the PQ generalized lasso problem
}
\keyword{internal}
