% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sp_chol_solve}
\alias{sp_chol_solve}
\alias{sp_chol_rnd_sim}
\alias{sp_chol_logdet}
\alias{sp_chol_edf}
\title{Cholesky solvers for sparse matrix operations}
\usage{
sp_chol_solve(R, r)

sp_chol_rnd_sim(R, r)

sp_chol_logdet(R)

sp_chol_edf(R, B, exact = FALSE, rank = 10, nsample = 100, seed = 1234)
}
\arguments{
\item{R}{upper-triangular factor of the Cholesky decomposition}

\item{r}{vector or matrix on the right-hand side of the linear system}

\item{B}{cross-product matrix}

\item{exact}{if \code{TRUE}, compute the exact edf}

\item{rank}{number of eigen-vectors to use}

\item{nsample}{number of random sample}

\item{seed}{random number generation seed}
}
\description{
\itemize{
\item \code{sp_chol_solve}: solve a linear system of equations
\item \code{sp_chol_rnd_sim}: generate a Gaussian random vector
\item \code{sp_chol_logdet}: log-determinant of a matrix
\item \code{sp_chol_edf}: Hutch++ estimator of the EDF
}
}
\keyword{internal}
