% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_enet_coord.R
\name{update_enet_coord_BL}
\alias{update_enet_coord_BL}
\title{Update beta via BL bound optimization for ridge logistic regression}
\usage{
update_enet_coord_BL(y, X, eta, w, beta, pL1, pL2, intercept, niter = 10)
}
\description{
Update beta via BL bound optimization for ridge logistic regression
}
\keyword{internal}
