% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_enet_coord.R
\name{update_enet_coord_PQ}
\alias{update_enet_coord_PQ}
\title{Update beta via PQ bound optimization for ridge logistic regression}
\usage{
update_enet_coord_PQ(
  y,
  X,
  eta,
  w,
  nu,
  s,
  beta,
  pL1,
  pL2,
  intercept,
  niter = 10
)
}
\description{
Update beta via PQ bound optimization for ridge logistic regression
}
\keyword{internal}
