% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_genlasso.R
\name{update_genlasso_BL}
\alias{update_genlasso_BL}
\title{Update beta via BL bound optimization for generalized-lasso logistic regression}
\usage{
update_genlasso_BL(beta, z, u, eta, prob, y, X, D, lambda, alpha, eps, ctr)
}
\description{
Update beta via BL bound optimization for generalized-lasso logistic regression
}
\keyword{internal}
