% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_genlasso.R
\name{update_genlasso_NR}
\alias{update_genlasso_NR}
\title{Update beta via NR quadratic approximation for generalized-lasso logistic regression}
\usage{
update_genlasso_NR(beta, z, u, eta, prob, y, X, w, D, lambda, alpha, eps, ctr)
}
\description{
Update beta via NR quadratic approximation for generalized-lasso logistic regression
}
\keyword{internal}
