% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_genlasso.R
\name{update_genlasso_PQ}
\alias{update_genlasso_PQ}
\title{Update beta via PQ bound optimization for generalized-lasso logistic regression}
\usage{
update_genlasso_PQ(
  beta,
  z,
  u,
  s,
  eta,
  y,
  X,
  w,
  nu,
  D,
  lambda,
  alpha,
  eps,
  phi,
  approx,
  ctr
)
}
\description{
Update beta via PQ bound optimization for generalized-lasso logistic regression
}
\keyword{internal}
