% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ridge.R
\name{update_ridge_BL}
\alias{update_ridge_BL}
\title{Update beta via BL bound optimization for ridge logistic regression}
\usage{
update_ridge_BL(y, X, eta, pL2, cholQ, solver = c("chol", "smw", "sparse"))
}
\description{
Update beta via BL bound optimization for ridge logistic regression
}
\keyword{internal}
