% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ridge.R
\name{update_ridge_NR}
\alias{update_ridge_NR}
\title{Update beta via NR quadratic approximation for ridge logistic regression}
\usage{
update_ridge_NR(
  y,
  X,
  eta,
  w,
  pL2,
  XPXt = NULL,
  XtWy = NULL,
  solver = c("chol", "smw", "sparse")
)
}
\description{
Update beta via NR quadratic approximation for ridge logistic regression
}
\keyword{internal}
