% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ridge.R
\name{update_ridge_PG}
\alias{update_ridge_PG}
\title{Update beta via PG bound optimization for ridge logistic regression}
\usage{
update_ridge_PG(
  y,
  X,
  eta,
  w,
  pL2,
  XtWy,
  XPXt = NULL,
  solver = c("chol", "smw", "sparse")
)
}
\description{
Update beta via PG bound optimization for ridge logistic regression
}
\keyword{internal}
