% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ridge.R
\name{update_ridge_PQ}
\alias{update_ridge_PQ}
\title{Update beta via PQ bound optimization for ridge logistic regression}
\usage{
update_ridge_PQ(
  y,
  X,
  eta,
  w,
  nu,
  s,
  pL2,
  beta,
  phi,
  XPXt = NULL,
  solver = c("chol", "smw", "admm", "sparse"),
  ctr = NULL
)
}
\description{
Update beta via PQ bound optimization for ridge logistic regression
}
\keyword{internal}
