% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_genlasso.R
\name{update_splasso_PG}
\alias{update_splasso_PG}
\title{Update beta via PG bound optimization for spatial-lasso logistic regression}
\usage{
update_splasso_PG(beta, z, u, y, X, w, D, lambda, alpha, eps, ctr)
}
\description{
Update beta via PG bound optimization for spatial-lasso logistic regression
}
\keyword{internal}
