% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prjgrad_large_n}
\alias{prjgrad_large_n}
\title{Projected gradient descent, optimized for large-n-small-p problems}
\usage{
prjgrad_large_n(u, Ny, NX, cholQ, alpha, thrs, maxiter)
}
\arguments{
\item{u}{(arma::vec(n)) initial solution}

\item{cholQ}{(arma::mat(p,p)) auxiliary Cholesky matrix}

\item{alpha}{(double) gradient step size}

\item{thrs}{(double) stopping threshold on relative absolute change}

\item{maxiter}{(int) maximum number of iterations}

\item{y_nu}{(arma::vec(n)) auxiliary vector}

\item{nu_x}{(arma::mat(n,p)) auxiliary matrix}
}
\value{
logistic penalized log likelihood
}
\description{
Projected gradient descent, optimized for large-n-small-p problems
}
