% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{set_ctr_admm}
\alias{set_ctr_admm}
\title{Set the control parameters of the ADMM algorithm}
\usage{
set_ctr_admm(
  rho = 1,
  tau = 0,
  gamma = 0,
  intercept = FALSE,
  spthr = 0.9,
  smw = FALSE,
  precondition = FALSE,
  objtol = 0.001,
  reltol = 0.001,
  abstol = 1e-04,
  maxiter = 100,
  verbose = FALSE,
  freq = 10
)
}
\description{
Set the control parameters of the ADMM algorithm
}
\keyword{internal}
