% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_genlasso.R
\name{solve_admm_genenet_PQ}
\alias{solve_admm_genenet_PQ}
\alias{solve_admm_genenet_WLS}
\title{Compute the generalized elastic-net solution using the ADMM algorithm}
\usage{
solve_admm_genenet_PQ(
  X,
  y,
  w,
  nu,
  D = diag(ncol(X)),
  d = numeric(nrow(D)),
  beta0 = NULL,
  z0 = NULL,
  u0 = NULL,
  lambda = 1,
  alpha = 0.99,
  eps = 1e-08,
  rho = 1,
  tau = 0,
  gamma = 0,
  intercept = TRUE,
  spthr = 0.9,
  smw = FALSE,
  precondition = FALSE,
  objtol = 1e-05,
  reltol = 0.01,
  abstol = 1e-04,
  maxiter = 1000,
  verbose = FALSE,
  freq = 10
)

solve_admm_genenet_WLS(
  X,
  y,
  w,
  D = diag(ncol(X)),
  d = numeric(nrow(D)),
  beta0 = NULL,
  z0 = NULL,
  u0 = NULL,
  lambda = 1,
  alpha = 0.99,
  eps = 1e-08,
  rho = 1,
  tau = 0,
  gamma = 0,
  intercept = TRUE,
  spthr = 0.9,
  smw = FALSE,
  precondition = FALSE,
  objtol = 1e-05,
  reltol = 0.01,
  abstol = 1e-04,
  maxiter = 1000,
  verbose = FALSE,
  freq = 10
)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{w}{L2 weight vector}

\item{nu}{L1 weight vector}

\item{D}{penalty matrix}

\item{beta0}{optional initial value for beta}

\item{z0}{optional initial value for the primal variables}

\item{u0}{optional initial value for the dual variables}

\item{lambda}{penalty parameter}

\item{alpha}{elastic-net mixing parameter}

\item{eps}{intercept specific penalty parameter}

\item{rho}{augmented Lagrangian penalty parameter}

\item{tau}{dual over-relaxation parameter}

\item{gamma}{proximal regularization parameter}

\item{intercept}{if \code{TRUE}, do not penalize per first column of \code{X}}

\item{spthr}{proportion of zero entries to transform \code{X} and \code{D} into sparse matrices}

\item{smw}{if \code{TRUE}, allows for the SMW solver in the \code{p>n} regime}

\item{precondition}{if \code{TRUE}, allows for diagonal preconditioning}

\item{objtol}{objective tolerance stopping criterion}

\item{reltol}{relative tolerance stopping criterion}

\item{abstol}{absolute tolerance stopping criterion}

\item{maxiter}{maximum number of iterations}

\item{verbose}{if \code{TRUE}, print the current optimization status}

\item{freq}{frequency of printing of the optimization status}
}
\value{
\itemize{
\item \code{beta}: estimated primal coefficients
\item \code{z}: estimated primal variable
\item \code{u}: estimated dual variables
\item \code{r}: primal residuals
\item \code{s}: dual residuals
\item \code{niter}: number of iterations to reach convergence
\item \code{trace}: data-frame collecting the optimization history
}
}
\description{
Compute the generalized elastic-net solution using the ADMM algorithm
}
