% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_genlasso.R
\name{solve_admm_ridge_PQ}
\alias{solve_admm_ridge_PQ}
\title{ADMM optimizer for the PQ generalized lasso problem}
\usage{
solve_admm_ridge_PQ(
  X,
  y,
  w,
  nu,
  pL2,
  beta0 = NULL,
  rho = 1,
  precondition = FALSE,
  objtol = 1e-05,
  reltol = 0.01,
  abstol = 1e-04,
  maxiter = 1000,
  history = FALSE
)
}
\arguments{
\item{X}{(real matrix) design matrix.}

\item{y}{(real vector) response vector.}

\item{w}{(real vector) weight vector.}

\item{nu}{(real vector) penalty weight vector.}

\item{pL2}{(real vector) ridge penalty vector.}

\item{beta0}{(real vector) initial values.}

\item{rho}{(real) augmented Lagrangian penalty parameter.}

\item{objtol}{(real) objective tolerance parameter.}

\item{reltol}{(real) relative tolerance parameter.}

\item{abstol}{(real) absolute tolerance parameter.}

\item{maxiter}{(int) maximum number of iterations.}

\item{lambda}{(real) generalized lasso penalty parameter.}
}
\value{
A list containing the estimated coefficients and the optimization history.
}
\description{
ADMM optimizer for the PQ generalized lasso problem
}
